+++ ragframework/document/__init__.py
@@ -3,7 +3,7 @@
 from ragframework.document.chunkers import FixedSizeChunker, RecursiveChunker, SentenceChunker
 
 from .html import HTMLLoader
-from .loaders import MarkdownLoader, PDFLoader, TextFileLoader
+from .loaders import DirectoryLoader, MarkdownLoader, PDFLoader, TextFileLoader
 from .tabular import CSVLoader, JSONLLoader
 
 __all__ = [
@@ -13,6 +13,7 @@ __all__ = [
     "CSVLoader",
     "JSONLLoader",
     "HTMLLoader",
+    "DirectoryLoader",
     "FixedSizeChunker",
     "RecursiveChunker",
     "SentenceChunker",